+++ inst/examples/simulate_zicreg_ex.R
@@ -1,5 +1,5 @@
 # Simulate some zip data
-n=1000
+n=300
 x = cbind(1, rnorm(n))
 z = cbind(1, rbeta(n, 4, 8))
 b = c(1, 2.2)